# Conflicts:
#	digiroad2-oracle/sql/insert_floating_road_addresses.sql
#	digiroad2-oracle/sql/insert_overlapping_road_addresses.sql
#	digiroad2-oracle/sql/insert_road_address_data.sql
#	digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/oracle/OracleDatabase.scala
#	digiroad2-viite/src/main/scala/fi/liikennevirasto/viite/dao/CalibrationPointDAO.scala
#	digiroad2-viite/src/main/scala/fi/liikennevirasto/viite/dao/LinearLocationDAO.scala
#	digiroad2-viite/src/main/scala/fi/liikennevirasto/viite/dao/ProjectLinkDAO.scala
#	digiroad2-viite/src/main/scala/fi/liikennevirasto/viite/dao/ProjectReservedPartDAO.scala
#	digiroad2-viite/src/main/scala/fi/liikennevirasto/viite/dao/RoadNameDAO.scala
#	digiroad2-viite/src/main/scala/fi/liikennevirasto/viite/dao/RoadwayDAO.scala
#	digiroad2-viite/src/main/scala/fi/liikennevirasto/viite/dao/UnaddressedRoadLinkDAO.scala
#	digiroad2-viite/src/main/scala/fi/liikennevirasto/viite/util/RoadAddressImporter.scala
#	digiroad2-viite/src/test/scala/fi/liikennevirasto/viite/ProjectServiceSpec.scala
#	digiroad2-viite/src/test/scala/fi/liikennevirasto/viite/RoadAddressServiceSpec.scala
#	digiroad2-viite/src/test/scala/fi/liikennevirasto/viite/util/package.scala
#	project/build.scala

+++ digiroad2-geo/src/main/scala/fi/liikennevirasto/digiroad2/linearasset/LinearAsset.scala
@@ -26,10 +26,10 @@ case class ValidityPeriod(val startHour: Int, val endHour: Int, val days: Validi
 
     if (endTotalMinutes > startTotalMinutes) {
       val duration = endTotalMinutes - startTotalMinutes
-      ((duration / 60).toInt, duration % 60)
+      ((duration / 60), duration % 60)
     } else {
       val duration = 1440 - startTotalMinutes + endTotalMinutes
-      ((duration / 60).toInt, duration % 60)
+      ((duration / 60), duration % 60)
     }
   }
